Move all middlewares from Traffic dir into Handler dir; rename HTTP `HandlerInterface` to `Middleware` interface.
