+++ src/Handler/Http/Middleware/DebugPage.php
@@ -2,22 +2,25 @@
 
 declare(strict_types=1);
 
-namespace Buggregator\Client\Traffic\Http;
+namespace Buggregator\Client\Handler\Http\Middleware;
 
+use Buggregator\Client\Handler\Http\Middleware;
 use Nyholm\Psr7\Response;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
-final class DebugPage implements HandlerInterface
+/**
+ * @internal
+ * @psalm-internal Buggregator\Client
+ */
+final class DebugPage implements Middleware
 {
-    public function handle(ServerRequestInterface $request, \Closure $next): ResponseInterface
+    public function handle(ServerRequestInterface $request, callable $next): ResponseInterface
     {
         if (\str_ends_with($request->getUri()->getPath(), '_debug')) {
             return new Response(
                 200,
-                [
-                    'Content-Type' => ['text/html; charset=UTF-8'],
-                ],
+                ['Content-Type' => ['text/html; charset=UTF-8']],
                 <<<'HTML'
                 <!doctype html>
                 <html>
