+++ src/Handler/Http/Middleware/RayRequestDump.php
@@ -2,16 +2,21 @@
 
 declare(strict_types=1);
 
-namespace Buggregator\Client\Traffic\Http;
+namespace Buggregator\Client\Handler\Http\Middleware;
 
+use Buggregator\Client\Handler\Http\Middleware;
 use Nyholm\Psr7\Response;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
-final class RayRequestDump implements HandlerInterface
+/**
+ * @internal
+ * @psalm-internal Buggregator\Client
+ */
+final class RayRequestDump implements Middleware
 {
 
-    public function handle(ServerRequestInterface $request, \Closure $next): ResponseInterface
+    public function handle(ServerRequestInterface $request, callable $next): ResponseInterface
     {
         if (\str_ends_with($request->getUri()->getPath(), '_availability_check')) {
             return new Response(400);
